+++ admin.html
@@ -240,7 +240,7 @@
           ><button class="btn buttonDetails my-5">Nuevo +</button></a
         >
       </section>
-      <section class="mx-4">
+      <section class="mx-4 table-responsive">
         <table class="table mx-auto text-light text-center">
           <thead>
             <tr class="element3">
@@ -252,34 +252,11 @@
               <th>Opciones</th>
             </tr>
           </thead>
-          <tbody>
-            <tr>
-              <th>123</th>
-              <th>FIFA ULtimate version</th>
-              <th>Disponible</th>
-              <th>
-                Lorem ipsum dolor sit amet consectetur adipisicing elit. Vel
-                repudiandae ea quod perferendis! Eligendi consequatur error odio
-                voluptatem esse culpa odit facere distinctio vel nihil!
-              </th>
-              <th>Si</th>
-              <th>simbolos</th>
-            </tr>
-            <tr>
-              <th>123</th>
-              <th>FIFA ULtimate version</th>
-              <th>Disponible</th>
-              <th>
-                Lorem ipsum dolor sit amet consectetur adipisicing elit. Quis
-                repudiandae, tempora perspiciatis repellat nemo ducimus?
-              </th>
-              <th>Si</th>
-              <th>simbolos</th>
-            </tr>
-          </tbody>
+          <tbody id="tablaJuegos"></tbody>
         </table>
       </section>
     </main>
+    <script src="js/loginGame.js" type="module"></script>
     <script src="js/bootstrap.min.js"></script>
     <script
       src="https://kit.fontawesome.com/7d2e342abf.js"
