+++ index.html
@@ -299,7 +299,7 @@
         </section>
       </section>
       <section class="container my-1">
-        <div class="row d-flex text-center">
+        <div class="row d-flex text-center" id="div123">
           <figure>
             <img
               src="/assets/joystick-line-pop-removebg-preview.png"
@@ -313,6 +313,7 @@
         </div>
       </section>
     </main>
+    <script src="js/login.js" type="module"></script>
     <script src="js/bootstrap.min.js"></script>
     <script
       src="https://kit.fontawesome.com/7d2e342abf.js"
