+++ js/login.js
@@ -3,12 +3,9 @@ import {
   validarEmail,
   validarRepetirContrasena,
   validacionGeneral,
-  validarFormularioJuego,
-  validarURL,
 } from "./validations.js";
 
 import { Usuario } from "./UsuarioClass.js";
-import { Game } from "./gameClass.js";
 
 let campoNombre = document.getElementById("Nombre");
 let campoApellido = document.getElementById("Apellido");
@@ -20,19 +17,9 @@ let campoFormLogin = document.getElementById("FormLogin");
 
 let FormInicioSesion = document.getElementById("FormInicioSesion");
 
-let campoNombreJuego = document.getElementById("NombreJuego");
-let campoCategoria = document.getElementById("Categoría");
-let campoDescripcion = document.getElementById("Descripcion");
-let campoPublicado = document.getElementById("Publicado");
-let campoURL = document.getElementById("URL");
-let campoFormJuego = document.getElementById("FormNewGame");
-let check;
-
-let juegoExistente = false;
 let usuarioExistente = false;
 
 let listaUsuario = JSON.parse(localStorage.getItem("Usuarios")) || [];
-let listaJuegos = JSON.parse(localStorage.getItem("Juegos")) || [];
 
 if (campoNombre) {
   campoNombre.addEventListener("blur", () => {
@@ -58,30 +45,6 @@ if (campoRepetirContrasena) {
   campoRepetirContrasena.addEventListener("blur", () => {
     validarRepetirContrasena(campoRepetirContrasena);
   });
-}
-if (campoNombreJuego) {
-  campoNombreJuego.addEventListener("blur", () => {
-    campoRequerido(campoNombreJuego);
-  });
-}
-if (campoCategoria) {
-  campoCategoria.addEventListener("blur", () => {
-    campoRequerido(campoCategoria);
-  });
-}
-if (campoDescripcion) {
-  campoDescripcion.addEventListener("blur", () => {
-    campoRequerido(campoDescripcion);
-  });
-}
-if (campoURL) {
-  campoURL.addEventListener("blur", () => {
-    validarURL(campoURL);
-  });
-}
-
-if (campoFormJuego) {
-  campoFormJuego.addEventListener("submit", agregarJuego);
 }
 
 if (campoFormLogin) {
@@ -126,15 +89,6 @@ function clearForm() {
   campoCheckboxRobot.className = "form-check";
 }
 
-function limpiarForm() {
-  campoFormJuego.reset();
-  campoNombreJuego.className = "form-control";
-  campoCategoria.className = "form-control";
-  campoDescripcion.className = "form-control";
-  campoURL.className = "form-control";
-  campoPublicado.className = "form-check-input";
-}
-
 function crearUsuario() {
   let usuarioNuevo = new Usuario(
     campoNombre.value,
@@ -149,10 +103,6 @@ function crearUsuario() {
 
 function guadarLocalStorage() {
   localStorage.setItem("Usuarios", JSON.stringify(listaUsuario));
-}
-
-function guardarJuegosEnLocalStorage() {
-  localStorage.setItem("Juegos", JSON.stringify(listaJuegos));
 }
 
 function InicioSesion(e) {
@@ -170,51 +120,3 @@ function InicioSesion(e) {
     });
   }
 }
-
-const numeroAleatorio = Math.floor(Math.random() * (99999 - 10000 + 1)) + 10000;
-
-function agregarJuego(event) {
-  event.preventDefault();
-  let botonAlerta = document.getElementById("btnAlerta");
-  if (campoPublicado) {
-    if (campoPublicado.checked) {
-      check = "Si";
-    } else {
-      check = "No";
-    }
-  }
-  if (
-    validarFormularioJuego(
-      campoNombreJuego,
-      campoCategoria,
-      campoDescripcion,
-      campoURL
-    )
-  ) {
-    botonAlerta.className = "alert alert-danger my-3 d-none";
-    if (!juegoExistente) {
-      crearJuego();
-    } else {
-      modificarJuego();
-    }
-  } else {
-    botonAlerta.className = "alert alert-danger my-3";
-  }
-}
-
-function crearJuego() {
-  let juegoNuevo = new Game(
-    numeroAleatorio,
-    campoNombreJuego.value,
-    campoCategoria.value,
-    campoDescripcion.value,
-    check,
-    campoURL.value
-  );
-  listaJuegos.push(juegoNuevo);
-  limpiarForm();
-  juegoExistente = false;
-  guardarJuegosEnLocalStorage();
-}
-
-function modificarJuego() {}
