+++ newGame.html
@@ -305,7 +305,7 @@
         </div>
       </section>
     </main>
-    <script src="js/login.js" type="module"></script>
+    <script src="js/loginGame.js" type="module"></script>
     <script src="js/bootstrap.min.js"></script>
     <script
       src="https://kit.fontawesome.com/7d2e342abf.js"
